Safari-611.1.10/LayoutTests/inspector/canvas/shaderProgram-add-remove-webgl.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/shaderProgram-utilities-webgl.js"></script>
<script>
function test() {
    let suite = initializeTestSuite("webgl");

    addSimpleTestCase("webgl");
    addParentCanvasRemovedTestCase("webgl");

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
    <p>Test that CanvasManager tracks creation and destruction of WebGL shader programs.</p>
</body>
<script>
    createProgram("webgl");
</script>
</html>
